+++ src/main/java/com/example/ajaxonboardingproject/nats/grpc/CinemaHallGrpcService.kt
@@ -6,9 +6,9 @@ import com.example.ajaxonboardingproject.CinemaHallServiceGrpc
 import com.example.ajaxonboardingproject.service.CinemaHallService
 import com.example.ajaxonboardingproject.service.proto.converter.CinemaHallConverter
 import io.grpc.stub.StreamObserver
-import org.springframework.stereotype.Controller
+import org.springframework.stereotype.Component
 
-@Controller
+@Component
 class CinemaHallGrpcService(
     private val cinemaHallService: CinemaHallService,
     private val cinemaHallConverter: CinemaHallConverter
