+++ src/test/java/com/example/ajaxonboardingproject/GrpcMovieServiceTest.kt
@@ -28,7 +28,7 @@ class GrpcMovieServiceTest {
     @BeforeEach
     fun startServer() {
         channel = ManagedChannelBuilder
-            .forAddress("localhost", 8096)
+            .forAddress("localhost", 8097)
             .usePlaintext()
             .build()
         stub = MovieServiceGrpc.newBlockingStub(channel)
