+++ src/test/java/com/example/ajaxonboardingproject/GrpcMovieSessionServiceTests.kt
@@ -32,7 +32,7 @@ class GrpcMovieSessionServiceTests {
     @BeforeEach
     fun startServer() {
         channel = ManagedChannelBuilder
-            .forAddress("localhost", 8096)
+            .forAddress("localhost", 8097)
             .usePlaintext()
             .build()
         stub = MovieSessionServiceGrpc.newBlockingStub(channel)
